+++ src/analysis/processing/qgsalgorithmcoverageunion.cpp
@@ -86,14 +86,14 @@ QVariantMap QgsCoverageUnionAlgorithm::processAlgorithm( const QVariantMap &para
   // we have to build up a list of geometries in advance
   QgsGeometryCollection collection;
 
-  const long count = source->featureCount();
+  const long long count = source->featureCount();
   if ( count >  0 )
   {
     collection.reserve( count );
   }
 
   const double step = count > 0 ? 100.0 / count : 1;
-  int current = 0;
+  long long current = 0;
 
   feedback->pushInfo( QObject::tr( "Collecting features" ) );
