+++ src/analysis/processing/qgsalgorithmmultiunion.cpp
@@ -193,8 +193,8 @@ bool QgsMultiUnionAlgorithm::makeUnion( const QgsFeatureSource &sourceA, const Q
   const QList<int> fieldIndicesA = QgsProcessingUtils::fieldNamesToIndices( QStringList(), sourceA.fields() );
   const QList<int> fieldIndicesB = QgsProcessingUtils::fieldNamesToIndices( QStringList(), sourceB.fields() );
 
-  long count = 0;
-  const long total = sourceA.featureCount() * 2 + sourceB.featureCount();
+  long long count = 0;
+  const long long total = sourceA.featureCount() * 2 + sourceB.featureCount();
 
   QgsOverlayUtils::intersection( sourceA, sourceB, sink, context, feedback, count, total, fieldIndicesA, fieldIndicesB );
   if ( feedback->isCanceled() )
